Batch and ROI mode validated via Model, view and controller pending
